Workaround for looting untransferrable items
For some reason, certain items seemingly cannot be loaded into vehicle
inventories (e.g., Soviet PK/RPK/sniper harnesses). When the user
selected the "loot" action, these failed, and the corpse would get
deleted as soon as this failure was encountered, without transferring
the remaining items off of the corprse.

This changes the logic so that a corpse will only get destroyed if the
transfer succeeds, and will continue to transfer items even if one
item is not successfully transferred.